Split fan.py to vendor-specific fan integrations
