+++ miio/integrations/fan/dmaker/__init__.py
@@ -1,2 +1,3 @@
 # flake8: noqa
+from .fan import FanP5
 from .fan_miot import Fan1C, FanMiot, FanP9, FanP10, FanP11
