+++ miio/integrations/fan/zhimi/__init__.py
@@ -1,2 +1,3 @@
 # flake8: noqa
-from .zhimi_miot import FanStatusZA5, FanZA5, OperationModeFanZA5
+from .fan import Fan, FanSA1, FanV2, FanZA1, FanZA4
+from .zhimi_miot import FanZA5
